requirements.txt
python-telegram-bot[job-queue]>=20.0
aiohttp>=3.8.0
matplotlib>=3.5.0
pandas>=1.3.0
asyncio>=3.4.3
python-dotenv>=0.19.0